requirements.txt
faiss_cpu==1.11.0
gradio==5.37.0
matplotlib==3.10.3
numpy==2.3.1
Pillow==11.3.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.66.1
transformers==4.53.2
